[core] moved particle type pair to common and removed the _internal namespace
